# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gymtime API",
    "description": "Gymtime health and invitation-only sign-in workflows.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/request-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "# Errors\nReturns safe input, origin, rate limit, or infrastructure errors.",
        "operationId": "request_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeResponse"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "# Errors\nReturns safe session or infrastructure errors.",
        "operationId": "session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "# Errors\nReturns safe input, origin, code, or infrastructure errors.",
        "operationId": "verify_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/teams/{token}": {
      "get": {
        "tags": [
          "public calendars"
        ],
        "summary": "# Errors\nReturns NotFound for invalid/rotated links, or safe storage errors. No internal fields are projected.",
        "operationId": "readPublicCalendar",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCalendarView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/schedule/actions": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "# Errors\nRequires current permissions, CSRF, an idempotency key and validated action data.",
        "operationId": "mutate",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "64 lowercase hex characters; reuse only for the same action",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeView"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/preview": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "# Errors\nRequires organizer permissions; validates template/closure data without saving it.",
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewView"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/{token}.ics": {
      "get": {
        "tags": [
          "public calendars"
        ],
        "summary": "# Errors\nReturns NotFound for invalid/rotated links, or safe calendar/storage errors.",
        "operationId": "feed",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/calendar": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeResponse": {
        "type": "object",
        "required": [
          "message",
          "resend_after_seconds"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "resend_after_seconds": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "issues"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldIssue"
            }
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "FieldIssue": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "OutcomeView": {
        "type": "object",
        "required": [
          "resources",
          "warnings"
        ],
        "properties": {
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceView"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PreviewItem": {
        "type": "object",
        "required": [
          "date"
        ],
        "properties": {
          "date": {
            "type": "string"
          },
          "end": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ErrorResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "space": {
            "type": [
              "string",
              "null"
            ]
          },
          "start": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PreviewView": {
        "type": "object",
        "required": [
          "occurrences",
          "affected_bookings",
          "affected_requests"
        ],
        "properties": {
          "affected_bookings": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "affected_requests": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "occurrences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreviewItem"
            }
          }
        }
      },
      "PublicCalendarView": {
        "type": "object",
        "required": [
          "team",
          "gym",
          "timezone",
          "webpage_url",
          "calendar_url",
          "subscription_url",
          "events"
        ],
        "properties": {
          "calendar_url": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicEventView"
            }
          },
          "gym": {
            "type": "string"
          },
          "subscription_url": {
            "type": "string"
          },
          "team": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "webpage_url": {
            "type": "string"
          }
        }
      },
      "PublicEventView": {
        "type": "object",
        "required": [
          "uid",
          "sequence",
          "starts_at",
          "ends_at",
          "activity",
          "space",
          "cancelled"
        ],
        "properties": {
          "activity": {
            "type": "string"
          },
          "cancelled": {
            "type": "boolean"
          },
          "ends_at": {
            "type": "integer",
            "format": "int64"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "space": {
            "type": "string"
          },
          "starts_at": {
            "type": "integer",
            "format": "int64"
          },
          "uid": {
            "type": "string"
          }
        }
      },
      "RequestCodeBody": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ResourceView": {
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "organizer",
          "csrf_token"
        ],
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "organizer": {
            "type": "boolean"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "VerifyCodeBody": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}